+++ CHANGELOG.md
@@ -7,7 +7,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 - Popup windows in main menu and flight screens showing initialization errors and unavailable API methods
 ### Changed
+- AscentClassic property `AutoPathPerc` was renamed to `AutoTurnPercent`
+- AscentClassic property `AutoPathSpeedFactor` was renamed to `AutoTurnSpeedFactor`
 - **All object instances** (modules, windows, controllers and operations) **are now permanent** until you reset KSP, so you can cache them and reuse them for multiple flights which may be useful if your program switches between vessels
+### Fixed
+- AscentClassic properties - `AutoPathPerc` and `AutoPathSpeedFactor`
 
 ## [0.5.1] - 2020-02-16
 ### Fixed